requirements.txt
Pillow>=10.0.0
requests>=2.31.0
python-barcode>=0.15.1
